# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistem Perwalian Mahasiswa STMIK Bandung",
        "description": "Dokumentasi REST API Enterprise untuk Sistem Perwalian Mahasiswa STMIK Bandung. Mengelola Autentikasi Sanctum, Data Dosen Wali, Data Mahasiswa, Pengajuan dan Review Perwalian, serta Rekapitulasi Laporan.",
        "contact": {
            "name": "Tim Capstone STMIK Bandung",
            "email": "admin@stmikbandung.ac.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Server API Lokal STMIK Bandung"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "Autentikasi Login Pengguna",
                "description": "Endpoint untuk login pengguna (Admin, Dosen, Mahasiswa) menggunakan Email dan Password.",
                "operationId": "7c0d984f3d009d07b4d7759429b1df11",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@stmikbandung.ac.id"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Admin123!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login Berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Login berhasil."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Kredensial atau Validasi Salah"
                    }
                }
            }
        },
        "/auth/register-admin": {
            "post": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "Registrasi Akun Admin Baru",
                "description": "Khusus registrasi administrator STMIK Bandung.",
                "operationId": "6ee0d628e01c116c4a99e241e589b1a5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Super Administrator"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "admin2@stmikbandung.ac.id"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "Admin123!"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "Admin123!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Admin Terdaftar"
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "Dapatkan Profil User Aktif",
                "description": "Mengembalikan data profil pengguna yang sedang login berdasarkan token Sanctum.",
                "operationId": "d5b50a2d7205555b662509be964c9256",
                "responses": {
                    "200": {
                        "description": "Profil ditemukan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/profile": {
            "put": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "Update Profil Pengguna",
                "description": "Memperbarui informasi nama, nomor hp, atau avatar akun user aktif.",
                "operationId": "009ba0c485de93eb37994e1bcc59f208",
                "responses": {
                    "200": {
                        "description": "Profil diperbarui"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/auth/logout": {
            "post": {
                "tags": [
                    "Autentikasi"
                ],
                "summary": "Logout Pengguna",
                "description": "Menghapus token akses Sanctum pengguna.",
                "operationId": "8f22b4b1d4d13235593868f409cf2395",
                "responses": {
                    "200": {
                        "description": "Logout Berhasil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/admin": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Dashboard Ringkasan Administrator",
                "description": "Menampilkan total statistik, grafik per semester, grafik status, grafik bulanan, dan recent activity.",
                "operationId": "7221f41767d1e5df3e80d46e8ba1dd7a",
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan statistik dashboard admin"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/dosen": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Dashboard Dosen Wali",
                "description": "Menampilkan total mahasiswa bimbingan, counter status perwalian, dan perwalian pending butuh tindakan.",
                "operationId": "be3c56f65757e749f7cca02ca6c60641",
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan statistik dashboard dosen"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/mahasiswa": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Dashboard Mahasiswa",
                "description": "Menampilkan total perwalian, status pengajuan aktif, dan timeline riwayat bimbingan.",
                "operationId": "90eadad825b8ddb152f44e6ff002635e",
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan statistik dashboard mahasiswa"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token Bearer Sanctum hasil autentikasi login (Contoh: 'Bearer {token}')",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}